Extract IO Utils in buildSrc
Issue gh-10455
